Create csv_to_json.py
Just to get the file up on Github. Here's the first version. It parses the CSV file into a Python dictionary. The key is the filer ID, associated with a value, which is another dictionary. This subdictionary contains the transaction code (A, B, or C), associated with a list of amounts for each donation.

The program does not currently convert this data to JSON.